Initial project version
- Includes the standard ExampleMod that comes with tModLoader.
I'm using this in my project because it makes modeling my code
after something that already works much easier.  Eventually I
will extrapolate my own code from the existing code so that my
mod can stand on its own.  I want to emphasize: the stuff in
this repo is NOT my own!  It is the property of blushiemagic,
the guy who made tModLoader.  I'm just adding to it.
- Includes an arbitrary .gitignore